【BUFF】添加了BUFF的的静态数据。黑夜里怪物的变化要使用BUFF实现。这个下次做吧~

+++ Assets/Scripts/Modules/Maze/Monster/Monster.cs
@@ -72,7 +72,7 @@ namespace GameLogic
 				return;
 			}
 
-			Script.UpdateHPBar(Info.HP, Info.GetAttribute(BattleAttribute.HP));
+            Script.UpdateHPBar(Info.HP, (int)Info.GetAttribute(BattleAttribute.HP));
 		}
 
 		public new MonsterRecord ToRecord()
